scripts/operator/tests/pyproject.toml
[tool.poetry]
name = "openshift-upgrade"
version = "0.1.0"
description = ""
authors = ["Josh Michielsen <github@mickey.dev>"]

[tool.poetry.dependencies]
python = "^3.6"
requests = "^2.23.0"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
